chore: Remove feature flag SubscriptionPlansPage
