+++ src/app/(proper_react)/(redesign)/(authenticated)/user/(dashboard)/dashboard/fix/data-broker-profiles/manual-remove/ManualRemoveView.tsx
@@ -131,13 +131,7 @@ export function ManualRemoveView(props: Props) {
                     elems: {
                       subscribe_link: (
                         <TelemetryLink
-                          href={
-                            props.enabledFeatureFlags.includes(
-                              "SubscriptionPlansPage",
-                            )
-                              ? "/subscription-plans"
-                              : "/user/dashboard/fix/data-broker-profiles/automatic-remove"
-                          }
+                          href="/subscription-plans"
                           eventData={{
                             link_id: "manual_removal_instructions_upsell",
                           }}
@@ -188,11 +182,7 @@ export function ManualRemoveView(props: Props) {
                 button_id: "manual_removal_upsell",
               },
             }}
-            href={
-              props.enabledFeatureFlags.includes("SubscriptionPlansPage")
-                ? "/subscription-plans"
-                : "/user/dashboard/fix/data-broker-profiles/automatic-remove"
-            }
+            href="/subscription-plans"
           >
             {l10n.getString(
               "fix-flow-data-broker-profiles-manual-remove-button-remove-for-me",
